Fix compilation issues and disable broken research modules
Several research modules have fundamental code issues that prevent compilation:
- WMMA fragment types undefined (frag_a, frag_b, etc.)
- Missing cooperative groups headers
- Missing BF16 headers
- CHECK_CUDA macro redefinition issues

Changes:
- Disable broken research modules (ncu, cuda, atomic, barrier, warp,
  mma, tensor_mem, wgmma, dp4a, fp8, graph, unified, multi_stream,
  mbarrier, coop, redux) until fixed
- Fix FP4/FP6 block scaling kernel (__half * float type error)
- Fix WMMA FP16 kernel (add proper fragment type aliases)

Note: Only memory, deep, advanced, and fp4 benchmarks currently compile
and run. The disabled modules need significant code fixes to work.

+++ src/sm_120/fp4_fp6_research_kernel.cu
@@ -228,8 +228,8 @@ __global__ void blockScalingKernel(const T* __restrict__ input,
         // Load scale for this block
         float scale = scales[block_id];
 
-        // Apply block scaling
-        output[idx] = input[idx] * scale;
+        // Apply block scaling (__half * float requires conversion)
+        output[idx] = __float2half(__half2float(input[idx]) * scale);
     }
 }
